Get Some meat on those bones
throwing error for some reason, if @Salamiboio could look into it that's be great.

(not finding m_GearScript in the script)

we can also move from OnControllableSpawned to every entity spawned if we want, sould allow the viewing of gear scripts in the editor window without having to launch the game.